[create] entity 모델링2

+++ server/src/entity/SectorWeighting.ts
@@ -15,7 +15,7 @@ export class SectorWeighting {
   @Column()
   sector: string
 
-  @Column({ type: 'decimal' })
+  @Column({ type: 'double' })
   percentage: number
 }
